backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.104.1
filelock==3.20.0
fsspec==2025.12.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.23.2
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6
nltk==3.9.2
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
portalocker==2.10.1
protobuf==6.33.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
qdrant-client==1.7.1
readchar==4.2.1
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==2.6.1
sentencepiece==0.2.1
shellingham==1.5.4
sniffio==1.3.1
socksio==1.0.0
specifyplus==0.0.19
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.43.3
triton==3.5.1
truststore==0.10.4
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.24.0.post1
